pyproject.toml
[project]
name = "autosentinel"
version = "0.1.0"
description = "AI-Native Observability & Self-Healing Agent System — Core Diagnostic Engine"
requires-python = ">=3.10"
dependencies = [
    "langgraph==1.1.9",
    "langgraph-checkpoint-postgres",
    "psycopg[binary]>=3.1",
    "anthropic>=0.40",
    "openai>=1.40",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "tenacity",
    "pydantic-settings",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["autosentinel*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["autosentinel"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 100
